refactor: split docs crate helpers
