+++ crates/ox_content_docs/src/extractor.rs
@@ -3,14 +3,17 @@
 mod jsdoc;
 mod jsdoc_tags;
 mod model;
+mod params;
+mod returns;
 mod tags;
+mod types;
 
 use oxc_allocator::Allocator;
 use oxc_ast::ast::{
     BindingPattern, Class, Comment, Declaration, ExportDefaultDeclarationKind, Expression,
     Function, Statement, TSEnumDeclaration, TSEnumMember, TSIndexSignature, TSIndexSignatureName,
     TSInterfaceDeclaration, TSSignature, TSType, TSTypeAliasDeclaration, TSTypeAnnotation,
-    TSTypeLiteral, TSTypeName, VariableDeclaration,
+    VariableDeclaration,
 };
 use oxc_ast_visit::{walk, Visit};
 use oxc_parser::Parser;
@@ -20,13 +23,13 @@ use std::path::Path;
 
 #[allow(unused_imports)]
 use crate::profile_span;
-use crate::string_builder::{join2, join3, join5, StringBuilder};
+use crate::string_builder::{join5, StringBuilder};
 
 use self::jsdoc::{build_jsdoc_cache, parse_jsdoc_payload, ParsedJsdoc, MODULE_MARKER_TAGS};
+use self::model::FunctionTypeMetadata;
 pub use self::model::{
     DocItem, DocItemKind, DocTag, ExtractError, ExtractResult, ParamDoc, TypeParamDoc,
 };
-use self::model::{FunctionTypeMetadata, ParsedParamTag};
 
 /// Documentation extractor.
 pub struct DocExtractor {
@@ -751,489 +754,6 @@ impl<'a> DocVisitor<'a> {
         }
 
         sig
-    }
-
-    /// Format a TypeScript type.
-    fn format_ts_type(&self, ts_type: &TSType) -> String {
-        match ts_type {
-            TSType::TSAnyKeyword(_) => "any".to_string(),
-            TSType::TSBooleanKeyword(_) => "boolean".to_string(),
-            TSType::TSNumberKeyword(_) => "number".to_string(),
-            TSType::TSStringKeyword(_) => "string".to_string(),
-            TSType::TSVoidKeyword(_) => "void".to_string(),
-            TSType::TSNullKeyword(_) => "null".to_string(),
-            TSType::TSUndefinedKeyword(_) => "undefined".to_string(),
-            TSType::TSNeverKeyword(_) => "never".to_string(),
-            TSType::TSBigIntKeyword(_) => "bigint".to_string(),
-            TSType::TSSymbolKeyword(_) => "symbol".to_string(),
-            TSType::TSObjectKeyword(_) => "object".to_string(),
-            TSType::TSUnknownKeyword(_) => "unknown".to_string(),
-            TSType::TSTypeReference(ref_type) => {
-                self.format_type_span(ref_type.span().start, ref_type.span().end)
-            }
-            TSType::TSArrayType(arr) => join2(&self.format_ts_type(&arr.element_type), "[]"),
-            TSType::TSTypeOperatorType(op) => {
-                let inner = self.format_ts_type(&op.type_annotation);
-                match op.operator {
-                    oxc_ast::ast::TSTypeOperatorOperator::Keyof => join2("keyof ", &inner),
-                    oxc_ast::ast::TSTypeOperatorOperator::Unique => join2("unique ", &inner),
-                    oxc_ast::ast::TSTypeOperatorOperator::Readonly => join2("readonly ", &inner),
-                }
-            }
-            TSType::TSUnionType(union) => {
-                let types: Vec<String> =
-                    union.types.iter().map(|t| self.format_ts_type(t)).collect();
-                types.join(" | ")
-            }
-            TSType::TSIntersectionType(inter) => {
-                let types: Vec<String> =
-                    inter.types.iter().map(|t| self.format_ts_type(t)).collect();
-                types.join(" & ")
-            }
-            TSType::TSFunctionType(func) => {
-                let params = self.format_type_formal_parameters(&func.params);
-                let ret = self.format_ts_type(&func.return_type.type_annotation);
-                let mut out = StringBuilder::with_capacity(params.len() + ret.len() + 6);
-                out.push_char('(');
-                out.push_str(&params);
-                out.push_str(") => ");
-                out.push_str(&ret);
-                out.into_string()
-            }
-            TSType::TSTypeLiteral(type_literal) => self.format_type_literal(type_literal),
-            TSType::TSParenthesizedType(paren) => {
-                join3("(", &self.format_ts_type(&paren.type_annotation), ")")
-            }
-            TSType::TSTupleType(tuple) => {
-                let types: Vec<String> = tuple
-                    .element_types
-                    .iter()
-                    .map(|t| self.format_ts_type(t.to_ts_type()))
-                    .collect();
-                join3("[", &types.join(", "), "]")
-            }
-            TSType::TSLiteralType(lit) => match &lit.literal {
-                oxc_ast::ast::TSLiteral::StringLiteral(s) => join3("\"", s.value.as_str(), "\""),
-                oxc_ast::ast::TSLiteral::NumericLiteral(n) => n
-                    .raw
-                    .as_ref()
-                    .map_or_else(|| n.value.to_string(), std::string::ToString::to_string),
-                oxc_ast::ast::TSLiteral::BooleanLiteral(b) => b.value.to_string(),
-                _ => "literal".to_string(),
-            },
-            _ => self.format_type_from_span(ts_type),
-        }
-    }
-
-    fn format_type_from_span(&self, ts_type: &TSType) -> String {
-        self.format_type_span(ts_type.span().start, ts_type.span().end)
-    }
-
-    fn format_span(&self, start: u32, end: u32) -> String {
-        self.slice(start, end).split_whitespace().collect::<Vec<_>>().join(" ").trim().to_string()
-    }
-
-    fn format_type_span(&self, start: u32, end: u32) -> String {
-        Self::collapse_type_annotation_text(&self.slice(start, end))
-    }
-
-    fn collapse_type_annotation_text(text: &str) -> String {
-        let text = text.trim();
-        if text.is_empty() {
-            return String::new();
-        }
-
-        let mut out = String::with_capacity(text.len());
-        let mut pending_space = false;
-        for ch in text.chars() {
-            if ch.is_whitespace() {
-                pending_space = !out.is_empty();
-                continue;
-            }
-
-            if pending_space {
-                if !matches!(out.chars().next_back(), Some('<')) && ch != '>' {
-                    out.push(' ');
-                }
-                pending_space = false;
-            }
-            out.push(ch);
-        }
-        out
-    }
-
-    fn property_key_name(key: &oxc_ast::ast::PropertyKey<'a>) -> Option<String> {
-        match key {
-            oxc_ast::ast::PropertyKey::StaticIdentifier(id) => Some(id.name.to_string()),
-            _ => None,
-        }
-    }
-
-    fn format_type_literal(&self, type_literal: &TSTypeLiteral<'a>) -> String {
-        let members = type_literal
-            .members
-            .iter()
-            .map(|member| self.format_type_literal_member(member))
-            .filter(|member| !member.is_empty())
-            .collect::<Vec<_>>();
-
-        if members.is_empty() {
-            "{}".to_string()
-        } else {
-            join3("{ ", &members.join("; "), " }")
-        }
-    }
-
-    fn format_type_literal_member(&self, member: &TSSignature<'a>) -> String {
-        match member {
-            TSSignature::TSPropertySignature(prop) => {
-                let Some(name) = Self::property_key_name(&prop.key) else {
-                    return self.format_span(prop.span.start, prop.span.end);
-                };
-                let type_annotation = prop.type_annotation.as_ref().map_or_else(
-                    || "unknown".to_string(),
-                    |t| self.format_ts_type(&t.type_annotation),
-                );
-                let mut out = StringBuilder::with_capacity(name.len() + type_annotation.len() + 16);
-                if prop.readonly {
-                    out.push_str("readonly ");
-                }
-                out.push_str(&name);
-                if prop.optional {
-                    out.push_char('?');
-                }
-                out.push_str(": ");
-                out.push_str(&type_annotation);
-                out.into_string()
-            }
-            TSSignature::TSMethodSignature(method) => {
-                let Some(name) = Self::property_key_name(&method.key) else {
-                    return self.format_span(method.span.start, method.span.end);
-                };
-                let params = self.format_type_formal_parameters(&method.params);
-                let return_type = method.return_type.as_ref().map_or_else(
-                    || "unknown".to_string(),
-                    |t| self.format_ts_type(&t.type_annotation),
-                );
-                let type_parameters =
-                    self.format_type_parameter_declaration(method.type_parameters.as_ref());
-                let mut out = StringBuilder::with_capacity(
-                    name.len() + type_parameters.len() + params.len() + return_type.len() + 8,
-                );
-                out.push_str(&name);
-                if method.optional {
-                    out.push_char('?');
-                }
-                out.push_str(&type_parameters);
-                out.push_char('(');
-                out.push_str(&params);
-                out.push_str("): ");
-                out.push_str(&return_type);
-                out.into_string()
-            }
-            TSSignature::TSIndexSignature(index_signature) => {
-                let Some(parameter) = index_signature.parameters.first() else {
-                    return self.format_span(index_signature.span.start, index_signature.span.end);
-                };
-                let (name, _, _) = self.format_index_signature_name(parameter);
-                let value_type =
-                    self.format_ts_type(&index_signature.type_annotation.type_annotation);
-                Self::format_index_signature(index_signature, &name, &value_type)
-            }
-            _ => self.format_span(member.span().start, member.span().end),
-        }
-    }
-
-    /// Format a TypeScript type name.
-    fn format_ts_type_name(name: &TSTypeName) -> String {
-        match name {
-            TSTypeName::IdentifierReference(id) => id.name.to_string(),
-            TSTypeName::QualifiedName(qn) => {
-                join3(&Self::format_ts_type_name(&qn.left), ".", qn.right.name.as_str())
-            }
-            TSTypeName::ThisExpression(_) => "this".to_string(),
-        }
-    }
-
-    fn extract_params_from_formals(
-        &self,
-        params: &oxc_ast::ast::FormalParameters<'a>,
-        tags: &[DocTag],
-    ) -> Vec<ParamDoc> {
-        // Parse the `@param`/`@arg`/`@argument` tags once, in source order,
-        // instead of re-filtering and re-parsing them for every formal
-        // parameter (the old `find_param_tag` did O(P*T) parses).
-        let parsed_param_tags: Vec<ParsedParamTag> = tags
-            .iter()
-            .filter(|tag| matches!(tag.tag.as_str(), "param" | "arg" | "argument"))
-            .filter_map(Self::parse_param_tag)
-            .collect();
-        let mut reserved_param_names = params
-            .items
-            .iter()
-            .filter_map(|param| Self::binding_pattern_identifier_name(&param.pattern))
-            .collect::<Vec<_>>();
-        if let Some(rest) = params.rest.as_ref() {
-            if let Some(name) = Self::binding_pattern_identifier_name(&rest.rest.argument) {
-                reserved_param_names.push(name);
-            }
-        }
-        let mut used_param_tag_indices = Vec::new();
-
-        let mut docs = Vec::with_capacity(params.items.len() + usize::from(params.rest.is_some()));
-
-        for param in &params.items {
-            let fallback_name = Self::binding_pattern_name(&param.pattern);
-            let tag_index = if Self::binding_pattern_is_destructured(&param.pattern) {
-                Self::find_destructured_param_tag_index(
-                    &parsed_param_tags,
-                    &used_param_tag_indices,
-                    &reserved_param_names,
-                )
-            } else {
-                Self::find_parsed_param_tag_index(&parsed_param_tags, &fallback_name)
-            };
-            let tag = tag_index.map(|index| &parsed_param_tags[index]);
-            if let Some(index) = tag_index {
-                used_param_tag_indices.push(index);
-            }
-            let name =
-                tag.and_then(Self::top_level_param_tag_name).unwrap_or(&fallback_name).to_string();
-            let default_value = self
-                .binding_pattern_default_value(&param.pattern)
-                .or_else(|| {
-                    param
-                        .initializer
-                        .as_ref()
-                        .map(|init| self.slice(init.span().start, init.span().end))
-                })
-                .or_else(|| tag.and_then(|tag| tag.default_value.clone()));
-
-            let type_annotation = param
-                .type_annotation
-                .as_ref()
-                .map(|t| self.format_ts_type(&t.type_annotation))
-                .or_else(|| tag.and_then(|tag| tag.type_annotation.clone()));
-
-            let optional =
-                param.optional || default_value.is_some() || tag.is_some_and(|tag| tag.optional);
-            let description = tag.and_then(|tag| tag.description.clone());
-
-            docs.push(ParamDoc {
-                name: name.clone(),
-                type_annotation,
-                optional,
-                default_value,
-                description,
-            });
-
-            if let Some(type_literal) = param
-                .type_annotation
-                .as_ref()
-                .and_then(|t| Self::type_literal_from_ts_type(&t.type_annotation))
-            {
-                docs.extend(self.extract_type_literal_param_members(
-                    &name,
-                    type_literal,
-                    &parsed_param_tags,
-                ));
-            }
-        }
-
-        if let Some(rest) = params.rest.as_ref() {
-            let name = Self::binding_pattern_name(&rest.rest.argument);
-            let tag_index = Self::find_parsed_param_tag_index(&parsed_param_tags, &name);
-            let tag = tag_index.map(|index| &parsed_param_tags[index]);
-            let type_annotation = rest
-                .type_annotation
-                .as_ref()
-                .map(|t| self.format_ts_type(&t.type_annotation))
-                .or_else(|| tag.and_then(|tag| tag.type_annotation.clone()));
-
-            docs.push(ParamDoc {
-                name: name.clone(),
-                type_annotation,
-                optional: tag.is_some_and(|tag| tag.optional),
-                default_value: tag.and_then(|tag| tag.default_value.clone()),
-                description: tag.and_then(|tag| tag.description.clone()),
-            });
-
-            if let Some(type_literal) = rest
-                .type_annotation
-                .as_ref()
-                .and_then(|t| Self::type_literal_from_ts_type(&t.type_annotation))
-            {
-                docs.extend(self.extract_type_literal_param_members(
-                    &name,
-                    type_literal,
-                    &parsed_param_tags,
-                ));
-            }
-        }
-
-        docs
-    }
-
-    fn type_literal_from_ts_type<'t>(ts_type: &'t TSType<'a>) -> Option<&'t TSTypeLiteral<'a>> {
-        match ts_type {
-            TSType::TSTypeLiteral(type_literal) => Some(type_literal),
-            TSType::TSParenthesizedType(paren) => {
-                Self::type_literal_from_ts_type(&paren.type_annotation)
-            }
-            _ => None,
-        }
-    }
-
-    fn extract_type_literal_param_members(
-        &self,
-        parent_name: &str,
-        type_literal: &TSTypeLiteral<'a>,
-        parsed_param_tags: &[ParsedParamTag],
-    ) -> Vec<ParamDoc> {
-        let mut params = Vec::new();
-        for member in &type_literal.members {
-            match member {
-                TSSignature::TSPropertySignature(prop) => {
-                    let Some(property_name) = Self::property_key_name(&prop.key) else {
-                        continue;
-                    };
-                    let lookup_name = join3(parent_name, ".", &property_name);
-                    let tag = Self::find_exact_parsed_param_tag(parsed_param_tags, &lookup_name);
-                    let optional = prop.optional || tag.is_some_and(|tag| tag.optional);
-                    let name =
-                        if optional { join2(&lookup_name, "?") } else { lookup_name.clone() };
-                    let type_annotation = prop
-                        .type_annotation
-                        .as_ref()
-                        .map(|t| self.format_ts_type(&t.type_annotation))
-                        .or_else(|| tag.and_then(|tag| tag.type_annotation.clone()));
-
-                    params.push(ParamDoc {
-                        name,
-                        type_annotation,
-                        optional,
-                        default_value: tag.and_then(|tag| tag.default_value.clone()),
-                        description: tag.and_then(|tag| tag.description.clone()),
-                    });
-                }
-                TSSignature::TSMethodSignature(method) => {
-                    let Some(method_name) = Self::property_key_name(&method.key) else {
-                        continue;
-                    };
-                    let lookup_name = join3(parent_name, ".", &method_name);
-                    let tag = Self::find_exact_parsed_param_tag(parsed_param_tags, &lookup_name);
-                    let optional = method.optional || tag.is_some_and(|tag| tag.optional);
-                    let name =
-                        if optional { join2(&lookup_name, "?") } else { lookup_name.clone() };
-                    let params_text = self.format_type_formal_parameters(&method.params);
-                    let return_type = method.return_type.as_ref().map_or_else(
-                        || "unknown".to_string(),
-                        |t| self.format_ts_type(&t.type_annotation),
-                    );
-                    let mut type_annotation =
-                        StringBuilder::with_capacity(params_text.len() + return_type.len() + 6);
-                    type_annotation.push_char('(');
-                    type_annotation.push_str(&params_text);
-                    type_annotation.push_str(") => ");
-                    type_annotation.push_str(&return_type);
-
-                    params.push(ParamDoc {
-                        name,
-                        type_annotation: Some(type_annotation.into_string()),
-                        optional,
-                        default_value: tag.and_then(|tag| tag.default_value.clone()),
-                        description: tag.and_then(|tag| tag.description.clone()),
-                    });
-                }
-                _ => {}
-            }
-        }
-        params
-    }
-
-    /// Extract parameters from a function.
-    fn extract_params(&self, func: &Function, tags: &[DocTag]) -> Vec<ParamDoc> {
-        self.extract_params_from_formals(&func.params, tags)
-    }
-
-    fn extract_return_type_from_annotation(
-        &self,
-        return_type: Option<&oxc_ast::ast::TSTypeAnnotation<'a>>,
-        tags: &[DocTag],
-    ) -> Option<String> {
-        return_type.map(|r| self.format_ts_type(&r.type_annotation)).or_else(|| {
-            tags.iter().find(|tag| tag.tag == "returns" || tag.tag == "return").and_then(|tag| {
-                let (type_annotation, description) = Self::parse_return_tag(tag);
-                type_annotation.or(description)
-            })
-        })
-    }
-
-    fn extract_return_from_annotation(
-        &self,
-        return_type: Option<&oxc_ast::ast::TSTypeAnnotation<'a>>,
-        tags: &[DocTag],
-    ) -> (Option<String>, Vec<DocItem>) {
-        if let Some(return_type) = return_type {
-            if let TSType::TSTypeLiteral(type_literal) = &return_type.type_annotation {
-                let members = self.extract_ts_signature_members(&type_literal.members);
-                let type_annotation = if members.is_empty() {
-                    self.format_ts_type(&return_type.type_annotation)
-                } else {
-                    "object".to_string()
-                };
-                return (Some(type_annotation), members);
-            }
-            return (Some(self.format_ts_type(&return_type.type_annotation)), Vec::new());
-        }
-
-        (self.extract_return_type_from_annotation(None, tags), Vec::new())
-    }
-
-    fn extract_return(&self, func: &Function, tags: &[DocTag]) -> (Option<String>, Vec<DocItem>) {
-        self.extract_return_from_annotation(func.return_type.as_ref().map(AsRef::as_ref), tags)
-    }
-
-    fn extract_function_type_metadata(
-        &self,
-        ts_type: &TSType<'a>,
-        tags: &[DocTag],
-    ) -> Option<FunctionTypeMetadata> {
-        match ts_type {
-            TSType::TSFunctionType(func) => {
-                let (return_type, return_members) =
-                    self.extract_return_from_annotation(Some(&func.return_type), tags);
-                Some(FunctionTypeMetadata {
-                    params: self.extract_params_from_formals(&func.params, tags),
-                    return_type,
-                    return_members,
-                    type_parameters: self.extract_type_parameters(func.type_parameters.as_ref()),
-                })
-            }
-            TSType::TSParenthesizedType(paren) => {
-                self.extract_function_type_metadata(&paren.type_annotation, tags)
-            }
-            TSType::TSTypeReference(ref_type) => {
-                let name = Self::type_reference_identifier_name(ref_type)?;
-                self.type_alias_function_metadata
-                    .get(&name)
-                    .map(FunctionTypeMetadata::as_reference_metadata)
-            }
-            TSType::TSIntersectionType(inter) => inter
-                .types
-                .iter()
-                .find_map(|ts_type| self.extract_function_type_metadata(ts_type, tags)),
-            _ => None,
-        }
-    }
-
-    fn type_reference_identifier_name(
-        ref_type: &oxc_ast::ast::TSTypeReference<'a>,
-    ) -> Option<String> {
-        match &ref_type.type_name {
-            TSTypeName::IdentifierReference(id) => Some(id.name.to_string()),
-            TSTypeName::QualifiedName(_) | TSTypeName::ThisExpression(_) => None,
-        }
     }
 
     /// Create a DocItem from a function.
