+++ crates/ox_content_docs/src/markdown/markdown_html.rs
@@ -6,17 +6,13 @@
 //! HTML structures (`<details>`, stats, member tables, prose blocks, …).
 
 use rustc_hash::FxHashSet;
-use std::sync::OnceLock;
-
-use regex::Regex;
 
 use super::{
-    cached_regex, clean_summary_text, collapse_type_annotation_whitespace, doc_kind_plural,
-    doc_page_href, effective_members_format, effective_parameters_format, entry_anchor, file_stem,
-    format_count_label, format_kind_label, generate_source_href, get_entry_badges, member_anchor,
-    normalize_signature, parse_example_block, process_doc_text, resolve_type_fragments, EntryStats,
-    ExampleBlock, MarkdownDisplayFormat, MarkdownDocsOptions, MarkdownLinkContext,
-    MarkdownPathStrategy, RegexCache, TypeFragment, DOC_KIND_ORDER,
+    clean_summary_text, doc_kind_plural, doc_page_href, effective_members_format,
+    effective_parameters_format, entry_anchor, file_stem, format_count_label, format_kind_label,
+    generate_source_href, get_entry_badges, member_anchor, normalize_signature,
+    parse_example_block, process_doc_text, EntryStats, ExampleBlock, MarkdownDisplayFormat,
+    MarkdownDocsOptions, MarkdownLinkContext, MarkdownPathStrategy, DOC_KIND_ORDER,
 };
 use crate::model::{
     ApiDocEntry, ApiDocMember, ApiDocModule, ApiDocTag, ApiParamDoc, ApiReturnDoc, ApiThrowsDoc,
@@ -25,378 +21,14 @@ use crate::model::{
 use crate::string_builder::{join3, StringBuilder};
 use rustc_hash::FxHashMap;
 
-fn escape_html(value: &str) -> String {
-    // Most inputs here are symbol names, type annotations, and kind labels,
-    // which usually contain no escapable HTML bytes. The early scan keeps that
-    // path to one pass plus `to_string()`, while the escaping path still does a
-    // single allocation sized near the input. This replaces chained
-    // `replace()` calls that performed five full-string passes and could
-    // allocate five intermediate strings for every rendered cell.
-    if !value.bytes().any(|b| matches!(b, b'&' | b'<' | b'>' | b'"' | b'\'')) {
-        return value.to_string();
-    }
-    let mut out = String::with_capacity(value.len() + 16);
-    for ch in value.chars() {
-        match ch {
-            '&' => out.push_str("&amp;"),
-            '<' => out.push_str("&lt;"),
-            '>' => out.push_str("&gt;"),
-            '"' => out.push_str("&quot;"),
-            '\'' => out.push_str("&#39;"),
-            _ => out.push(ch),
-        }
-    }
-    out
-}
+mod blocks;
+mod inline;
 
-fn render_doc_inline_html(text: &str, context: Option<&MarkdownLinkContext<'_>>) -> String {
-    render_inline_html(&process_doc_text(text, context))
-}
-
-fn replace_line_breaks_html(html: String) -> String {
-    if html.contains('\n') {
-        html.replace('\n', "<br>")
-    } else {
-        html
-    }
-}
-
-fn render_inline_html(text: &str) -> String {
-    static TOKEN_RE: RegexCache = OnceLock::new();
-
-    let Some(token_re) = cached_regex(
-        &TOKEN_RE,
-        r"`([^`]+)`|\[([^\]]+)\]\(([^)]+)\)|\*\*([^*]+)\*\*|__([^_]+)__|\*([^*]+)\*|_([^_]+)_",
-    ) else {
-        return replace_line_breaks_html(escape_html(text));
-    };
-    // Check for a token before allocating the output buffer or entering the
-    // capture iterator. Plain prose dominates generated descriptions, and for
-    // that case escaping plus newline replacement is enough.
-    if !token_re.is_match(text) {
-        return replace_line_breaks_html(escape_html(text));
-    }
-
-    let mut html = String::new();
-    let mut last_index = 0;
-
-    for captures in token_re.captures_iter(text) {
-        let Some(mat) = captures.get(0) else {
-            continue;
-        };
-        html.push_str(&escape_html(&text[last_index..mat.start()]));
-
-        if let Some(code) = captures.get(1) {
-            html.push_str("<code>");
-            html.push_str(&escape_html(code.as_str()));
-            html.push_str("</code>");
-        } else if let (Some(label), Some(href)) = (captures.get(2), captures.get(3)) {
-            html.push_str("<a href=\"");
-            html.push_str(&escape_html(href.as_str()));
-            html.push_str("\">");
-            html.push_str(&render_inline_html(label.as_str()));
-            html.push_str("</a>");
-        } else if let Some(strong) = captures.get(4).or_else(|| captures.get(5)) {
-            html.push_str("<strong>");
-            html.push_str(&render_inline_html(strong.as_str()));
-            html.push_str("</strong>");
-        } else if let Some(emphasis) = captures.get(6).or_else(|| captures.get(7)) {
-            html.push_str("<em>");
-            html.push_str(&render_inline_html(emphasis.as_str()));
-            html.push_str("</em>");
-        }
-
-        last_index = mat.end();
-    }
-
-    html.push_str(&escape_html(&text[last_index..]));
-    replace_line_breaks_html(html)
-}
-
-// The four block-start regexes are cached once and shared between the
-// value-returning helpers, which need captures, and `is_markdown_block_start`,
-// which only needs a boolean and therefore uses allocation-free `is_match`.
-// This keeps list/paragraph continuation checks from rebuilding regexes or
-// allocating capture groups on every line.
-fn fence_re() -> Option<&'static Regex> {
-    static FENCE_RE: RegexCache = OnceLock::new();
-    cached_regex(&FENCE_RE, r"^```([\w-]+)?\s*$")
-}
-
-fn heading_re() -> Option<&'static Regex> {
-    static HEADING_RE: RegexCache = OnceLock::new();
-    cached_regex(&HEADING_RE, r"^(#{1,6})\s+(.*)$")
-}
-
-fn ordered_re() -> Option<&'static Regex> {
-    static ORDERED_RE: RegexCache = OnceLock::new();
-    cached_regex(&ORDERED_RE, r"^\d+\.\s+(.*)$")
-}
-
-fn unordered_re() -> Option<&'static Regex> {
-    static UNORDERED_RE: RegexCache = OnceLock::new();
-    cached_regex(&UNORDERED_RE, r"^[-*+]\s+(.*)$")
-}
-
-fn is_fence_start(line: &str) -> Option<String> {
-    fence_re()?
-        .captures(line.trim())
-        .map(|captures| captures.get(1).map_or("text", |value| value.as_str()).to_string())
-}
-
-fn heading_match(line: &str) -> Option<(usize, String)> {
-    heading_re()?.captures(line.trim()).map(|captures| {
-        (
-            captures.get(1).map_or(1, |value| value.as_str().len()).min(6),
-            captures.get(2).map_or("", |value| value.as_str()).trim().to_string(),
-        )
-    })
-}
-
-fn ordered_list_item(line: &str) -> Option<String> {
-    ordered_re()?
-        .captures(line.trim())
-        .and_then(|captures| captures.get(1).map(|value| value.as_str().to_string()))
-}
-
-fn unordered_list_item(line: &str) -> Option<String> {
-    unordered_re()?
-        .captures(line.trim())
-        .and_then(|captures| captures.get(1).map(|value| value.as_str().to_string()))
-}
-
-fn is_markdown_block_start(line: &str) -> bool {
-    // Only a boolean is needed, so use `is_match` (no capture allocation) and
-    // trim once. The regexes are `^…$`-anchored, so `is_match(trimmed)` is true
-    // exactly when the corresponding `captures(trimmed)` was `Some`.
-    let trimmed = line.trim();
-    fence_re().is_some_and(|re| re.is_match(trimmed))
-        || heading_re().is_some_and(|re| re.is_match(trimmed))
-        || ordered_re().is_some_and(|re| re.is_match(trimmed))
-        || unordered_re().is_some_and(|re| re.is_match(trimmed))
-}
-
-fn render_markdown_blocks_html(text: &str) -> String {
-    // This renderer handles the small Markdown subset embedded in generated
-    // API descriptions. It walks the line slice once and emits blocks as soon
-    // as they are recognized. Continuation checks use cached regexes and the
-    // inline renderer's own token precheck so ordinary paragraphs do not pay
-    // for full Markdown parsing.
-    static ORDERED_CONTINUATION_RE: RegexCache = OnceLock::new();
-    static UNORDERED_CONTINUATION_RE: RegexCache = OnceLock::new();
-
-    let lines: Vec<&str> =
-        text.split('\n').map(|line| line.strip_suffix('\r').unwrap_or(line)).collect();
-    let mut blocks = Vec::new();
-    let mut index = 0;
-    let ordered_continuation_re = cached_regex(&ORDERED_CONTINUATION_RE, r"^ {0,1}\d+\.\s+");
-    let unordered_continuation_re = cached_regex(&UNORDERED_CONTINUATION_RE, r"^[-*+]\s+");
-
-    while index < lines.len() {
-        let line = lines[index];
-        let trimmed = line.trim();
-
-        if trimmed.is_empty() {
-            index += 1;
-            continue;
-        }
-
-        if let Some(language) = is_fence_start(line) {
-            let mut code_lines = Vec::new();
-            index += 1;
-
-            while index < lines.len() && !lines[index].trim().starts_with("```") {
-                code_lines.push(lines[index]);
-                index += 1;
-            }
-
-            if index < lines.len() {
-                index += 1;
-            }
-
-            blocks.push(render_code_block_html(&code_lines.join("\n"), &language));
-            continue;
-        }
-
-        if let Some((level, content)) = heading_match(line) {
-            let content = render_inline_html(&content);
-            let mut block = StringBuilder::with_capacity(content.len() + 9);
-            block.push_str("<h");
-            block.push_usize(level);
-            block.push_char('>');
-            block.push_str(&content);
-            block.push_str("</h");
-            block.push_usize(level);
-            block.push_char('>');
-            blocks.push(block.into_string());
-            index += 1;
-            continue;
-        }
-
-        if let Some(first_item) = ordered_list_item(line) {
-            let mut items = Vec::new();
-            let mut current = Some(first_item);
-
-            while index < lines.len() {
-                let Some(item_text) = current.take().or_else(|| ordered_list_item(lines[index]))
-                else {
-                    break;
-                };
-
-                let mut item_lines = Vec::new();
-                item_lines.push(String::from(item_text.trim()));
-                index += 1;
-
-                while index < lines.len() {
-                    let continuation = lines[index];
-                    let continuation_trimmed = continuation.trim();
-
-                    if continuation_trimmed.is_empty()
-                        || is_markdown_block_start(continuation)
-                        || ordered_continuation_re
-                            .is_some_and(|re| re.is_match(continuation_trimmed))
-                    {
-                        break;
-                    }
-
-                    item_lines.push(continuation_trimmed.to_string());
-                    index += 1;
-                }
-
-                items.push(join3("<li>", &render_inline_html(&item_lines.join(" ")), "</li>"));
-
-                if index < lines.len() && lines[index].trim().is_empty() {
-                    break;
-                }
-            }
-
-            blocks.push(join3("<ol>\n", &items.join("\n"), "\n</ol>"));
-            continue;
-        }
-
-        if let Some(first_item) = unordered_list_item(line) {
-            let mut items = Vec::new();
-            let mut current = Some(first_item);
-
-            while index < lines.len() {
-                let Some(item_text) = current.take().or_else(|| unordered_list_item(lines[index]))
-                else {
-                    break;
-                };
-
-                let mut item_lines = Vec::new();
-                item_lines.push(String::from(item_text.trim()));
-                index += 1;
-
-                while index < lines.len() {
-                    let continuation = lines[index];
-                    let continuation_trimmed = continuation.trim();
-
-                    if continuation_trimmed.is_empty()
-                        || is_markdown_block_start(continuation)
-                        || unordered_continuation_re
-                            .is_some_and(|re| re.is_match(continuation_trimmed))
-                    {
-                        break;
-                    }
-
-                    item_lines.push(continuation_trimmed.to_string());
-                    index += 1;
-                }
-
-                items.push(join3("<li>", &render_inline_html(&item_lines.join(" ")), "</li>"));
-
-                if index < lines.len() && lines[index].trim().is_empty() {
-                    break;
-                }
-            }
-
-            blocks.push(join3("<ul>\n", &items.join("\n"), "\n</ul>"));
-            continue;
-        }
-
-        let mut paragraph_lines = Vec::new();
-        paragraph_lines.push(String::from(trimmed));
-        index += 1;
-
-        while index < lines.len() {
-            let next_line = lines[index];
-            let next_trimmed = next_line.trim();
-
-            if next_trimmed.is_empty() || is_markdown_block_start(next_line) {
-                break;
-            }
-
-            paragraph_lines.push(next_trimmed.to_string());
-            index += 1;
-        }
-
-        blocks.push(join3("<p>", &render_inline_html(&paragraph_lines.join(" ")), "</p>"));
-    }
-
-    join3("<div class=\"ox-api-entry__prose\">\n", &blocks.join("\n"), "\n</div>")
-}
-
-fn render_code_block_html(code: &str, language: &str) -> String {
-    let code = escape_html(code);
-    let mut out = StringBuilder::with_capacity(
-        "<pre><code class=\"language-\"></code></pre>".len() + language.len() + code.len(),
-    );
-    out.push_str("<pre><code class=\"language-");
-    out.push_str(language);
-    out.push_str("\">");
-    out.push_str(&code);
-    out.push_str("</code></pre>");
-    out.into_string()
-}
-
-fn render_highlighted_inline_code_html(code: &str, class_name: &str, language: &str) -> String {
-    let class_name = escape_html(class_name);
-    let code = escape_html(code);
-    let mut out = StringBuilder::with_capacity(
-        "<code class=\" language-\"></code>".len() + class_name.len() + language.len() + code.len(),
-    );
-    out.push_str("<code class=\"");
-    out.push_str(&class_name);
-    out.push_str(" language-");
-    out.push_str(language);
-    out.push_str("\">");
-    out.push_str(&code);
-    out.push_str("</code>");
-    out.into_string()
-}
-
-/// Inner HTML for a TypeScript type annotation: escaped text with `<a>` anchors
-/// for known symbols. Type annotations are always rendered as inline code, so
-/// multiline generic formatting is collapsed before link resolution.
-fn render_type_inner_html(
-    value: &str,
-    context: Option<&MarkdownLinkContext<'_>>,
-    skip: &FxHashSet<&str>,
-) -> String {
-    let value = collapse_type_annotation_whitespace(value);
-    match resolve_type_fragments(&value, context, skip) {
-        None => escape_html(&value),
-        Some(fragments) => {
-            let mut out = String::new();
-            for fragment in fragments {
-                match fragment {
-                    TypeFragment::Text(text) | TypeFragment::Code(text) => {
-                        out.push_str(&escape_html(&text));
-                    }
-                    TypeFragment::Link { name, href } => {
-                        out.push_str("<a href=\"");
-                        out.push_str(&escape_html(&href));
-                        out.push_str("\">");
-                        out.push_str(&escape_html(&name));
-                        out.push_str("</a>");
-                    }
-                }
-            }
-            out
-        }
-    }
-}
+use blocks::render_markdown_blocks_html;
+use inline::{
+    escape_html, render_code_block_html, render_doc_inline_html,
+    render_highlighted_inline_code_html, render_inline_html, render_type_inner_html,
+};
 
 pub(super) fn render_details_controls_html(target_selector: &str) -> String {
     let mut out = StringBuilder::with_capacity(260 + target_selector.len());
